@@ -1,7 +1,7 @@
 import {Component, Input} from '@angular/core';
 import {AngularFireDatabase} from 'angularfire2/database';
 import {AngularFireAuth} from 'angularfire2/auth';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import * as firebase from 'firebase/app';
 import {User} from 'firebase/app';
 import {Video} from './video-item/video';
@@ -9,7 +9,7 @@ import {Video} from './video-item/video';
 @Component({
     selector: 'app-root',
     templateUrl: './app.component.html',
-    styleUrls: ['./app.component.css']
+    styleUrls: ['./app.component.css'],
 })
 export class AppComponent {
     @Input() title: string;
@@ -21,10 +21,12 @@ export class AppComponent {
         this.user = this.afAuth.authState;
         this.afAuth.authState.subscribe(auth => {
             if (auth) {
-                af.list('/videos').subscribe(vids => {
-                    this.videos = vids;
-                    this.filterVideos();
-                });
+                af.list('/videos')
+                    .valueChanges()
+                    .subscribe(vids => {
+                        this.videos = <Video[]>vids;
+                        this.filterVideos();
+                    });
             } else {
                 this.videos = null;
                 this.filteredVideos = null;
