@@ -6,7 +6,15 @@ import {AngularFireModule} from 'angularfire2';
 import {AngularFireDatabaseModule} from 'angularfire2/database';
 import {AngularFireAuthModule} from 'angularfire2/auth';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
-import {MdButtonModule, MdCardModule, MdIconModule, MdInputModule, MdMenuModule, MdToolbarModule, MdTooltipModule} from '@angular/material';
+import {
+    MatButtonModule,
+    MatCardModule,
+    MatIconModule,
+    MatInputModule,
+    MatMenuModule,
+    MatToolbarModule,
+    MatTooltipModule,
+} from '@angular/material';
 import {FlexLayoutModule} from '@angular/flex-layout';
 
 import {AppComponent} from './app.component';
@@ -35,13 +43,13 @@ export const firebaseConfig = {
         AngularFireAuthModule,
         BrowserAnimationsModule,
         FlexLayoutModule,
-        MdButtonModule,
-        MdMenuModule,
-        MdCardModule,
-        MdToolbarModule,
-        MdTooltipModule,
-        MdIconModule,
-        MdInputModule
+        MatButtonModule,
+        MatMenuModule,
+        MatCardModule,
+        MatToolbarModule,
+        MatTooltipModule,
+        MatIconModule,
+        MatInputModule
     ],
     providers: [],
     bootstrap: [AppComponent]
